Move project pages to posts, use excerpt on articles
When I realised that each post can have a custom permalink set in its YAML, move
all the projects from a tree of index.html files to entries in the _posts dir with
project page layout and permalink set in the YAML.

Intermingle updated projects and blog entries on the home page.

Use the excerpt field with a "more" marker in the blog entries.